docker/apps/tinyolly-otlp-receiver/requirements.txt
grpcio>=1.60.0
protobuf>=5.0.0
opentelemetry-proto>=1.24.0
uvloop>=0.19.0
opentelemetry-distro>=0.42b0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-logging>=0.42b0
